GUI has SOME scripts attached to SOME buttons.
Point me in the direction of the other scripts, when you can. Please by the way, be thorough with your documentation going forward. I'm trying with mine and in the future I will be more in-depth with what each function does so that there's no confusion with my script. Please do the same because I need to be able to place these in the proper sections.